# openapi.yaml
openapi: 3.0.3
info:
  title: MVTMS REST API
  description: |
    OpenAPI specification for the MVTMS REST API v1. Authentication is via
    a signed Bearer token (`Authorization: Bearer <token>`) or an existing
    session cookie. The token endpoints are public; all other `/api/v1/*`
    routes require a valid principal plus Casbin RBAC permissions.
  version: "1.0.0"
  contact:
    name: MVTMS Support
    url: https://avandab.com
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/trips:
    get:
      tags: [Trips]
      summary: List trips
      description: Requires `trips:read` permission.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/StatusParam'
      responses:
        '200':
          description: Paginated list of trips.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Trips]
      summary: Create a trip
      description: Requires `trips:create` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripRequest'
      responses:
        '201':
          description: Trip created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IDResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed API token returned by `POST /api/v1/auth/token`.
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Page number (1-based).
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 10
      description: Items per page.
    SearchParam:
      name: search
      in: query
      schema:
        type: string
      description: Free-text search filter.
    StatusParam:
      name: status
      in: query
      schema:
        type: string
      description: Status filter.
  schemas:
...
    IDResponse:
      type: object
      properties:
        id:
          type: string
...
    CreateTripRequest:
      type: object
      required: [route_id, departure_time]
      properties:
        booking_id:
          type: string
          nullable: true
        route_id:
          type: string
        departure_time:
          type: string
          format: date-time
        remarks:
          type: string
    Trip:
      type: object
      properties:
        id:
          type: string
        trip_number:
          type: string
        booking_id:
          type: string
          nullable: true
        driver_id:
          type: string
          nullable: true
        driver_display_id:
          type: string
        driver_first_name:
          type: string
        driver_last_name:
          type: string
        vehicle_id:
          type: string
          nullable: true
        vehicle_registration_number:
          type: string
        vehicle_number:
          type: string
        route_id:
          type: string
        route_source:
          type: string
        route_destination:
          type: string
        departure_time:
          type: string
          format: date-time
        arrival_time:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [draft, scheduled, assigned, started, reached_pickup, in_transit, delivered, completed, cancelled]
        remarks:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        reached_pickup_at:
          type: string
          format: date-time
          nullable: true
        in_transit_at:
          type: string
          format: date-time
          nullable: true
        delivered_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
    TripList:
      type: object
      properties:
        trips:
          type: array
          items:
            $ref: '#/components/schemas/Trip'
        total:
          type: integer
          format: int64
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or invalid credentials.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient RBAC permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'